primer intento de submit

+++ Aplicacion/main.py
@@ -2,8 +2,7 @@ import time
 
 from minikube.mk import minikube
 from parsers.argparser import argparser
-from parsers.jsoncontroller import JsonController
-from kubernetes_control.kubernetes_control import kubernetes_control
+from spark.spark import spark
 import yaml
 from parsers.jsoncontroller import JsonController
 
@@ -30,35 +29,51 @@ if Debug:
 var = vars(args)
 mk = minikube(var.get('Verbose'))
 j = JsonController()
-if var.get('File'):
+sparkHome = j.get_object('spark-home')
+if var.get('sparkHome'):
+    j.add_object('spark-home', 'path', var.get('sparkHome'))
+    exit(0)
 
-    if var.get('fs'):
-        mk_template_path = ('./templates/default-mk.yaml')
-        spark_template = parser_yaml(var.get('File'))
-    else:
-        mk_template_path = var.get('File')
-
-    j.add_object('template-path', 'path', mk_template_path)
-
-    mk_template = parser_yaml(mk_template_path)
-    mk.start(mk_template['MiniKube'])
-    time.sleep(5)
-    kb = kubernetes_control(mk_template['MiniKube']['namespace'])
-    print(kb.master_ip())
-
-if var.get('cluster'):
-    status = var.get('cluster')
-    print(status)
-
-    if status == 'stop' or status == 'pause' \
-            or status == 'unpause':
-        mk.status(status)
-
-    elif status == 'restart':
-        mk.start(restart=True)
-
-    elif status == 'delete':
-        print('Deleting...')
-        mk.delete()
+if sparkHome:
+    print(sparkHome)
+    print(var)
+    if var.get('test'):
+        sp = spark()
+        sp.submit()
 
 
+    if var.get('File'):
+
+        if var.get('fs'):
+            yaml_template = parser_yaml('./templates/default-mk.yaml')
+            mk_template = yaml_template['MiniKube']
+            spark_template = parser_yaml(var.get('File'))['Spark']
+
+        else:
+            full_template = var.get('File')
+            mk_template = full_template['MiniKube']
+            spark_template = full_template['Spark']
+
+        mk.start(mk_template)
+        #sp = spark(spark_template)
+        #sp.submit()
+
+
+    if var.get('cluster'):
+        status = var.get('cluster')
+        print(status)
+
+        if status == 'stop' or status == 'pause' \
+                or status == 'unpause':
+            mk.status(status)
+
+        elif status == 'restart':
+            mk.start(restart=True)
+
+        elif status == 'delete':
+            print('Deleting...')
+            mk.delete()
+
+
+else:
+    print('First you need to get the path for spark')
